+++ ts/object/src/index.ts
@@ -2,6 +2,7 @@
 export * from './lib/decorators/after'
 export * from './lib/decorators/before'
 export * from './lib/decorators/build'
+export * from './lib/decorators/coerce'
 export * from './lib/decorators/delegate'
 export * from './lib/decorators/nonenumerable'
 export * from './lib/decorators/include'
@@ -14,6 +15,7 @@ export * from './lib/decorators/stack'
 
 export * from './lib/descriptors'
 
-
+export * from './lib/inflate'
 export * from './lib/deflate'
 export * from './lib/meta'
+export * from './lib/types'